Include all the logic on NTMainPanel
